renamed Chart component for network & country
