+++ components/network/ChartNetwork.js
@@ -1,6 +1,6 @@
 import React, { useMemo } from 'react'
+import { useIntl } from 'react-intl'
 import { useRouter } from 'next/router'
-import dayjs from 'services/dayjs'
 import { Heading, Box, Flex } from 'ooni-components'
 import useSWR from 'swr'
 import GridChart, { prepareDataForGridChart } from 'components/aggregation/mat/GridChart'
@@ -13,9 +13,10 @@ const swrOptions = {
   dedupingInterval: 10 * 60 * 1000,
 }
 
-const Chart = React.memo(function Chart({testName, testGroup = null, title, queryParams = {}}) {
+const ChartNetwork = React.memo(function Chart({testName, testGroup = null, title, queryParams = {}}) {
+  const intl = useIntl()
   const router = useRouter()
-  const { query: { countryCode} } = router
+  const { query: {since, until, asn} } = router
 
   const name = testName || testGroup.name
 
@@ -26,15 +27,14 @@ const Chart = React.memo(function Chart({testName, testGroup = null, title, quer
 
   const query = useMemo(() => ({
     ...params,
-    probe_cc: countryCode,
-    since: dayjs.utc().subtract(30, 'day').format('YYYY-MM-DD'),
-    until: dayjs.utc().add(1, 'day').format('YYYY-MM-DD'),
+    probe_asn: asn,
+    since: since,
+    until: until,
     ...testName && {test_name: testName}
-  }), [countryCode, params, testName])
+  }), [since, until, asn, params, testName])
 
   const apiQuery = useMemo(() => {
     const qs = new URLSearchParams(query).toString()
-    console.log(qs)
     return qs
   }, [query])
 
@@ -52,9 +52,9 @@ const Chart = React.memo(function Chart({testName, testGroup = null, title, quer
     }
     let chartData = testGroup ? data : data.data
     const graphQuery = testGroup ? {...query, axis_y: name} : query
-    const [reshapedData, rowKeys, rowLabels] = prepareDataForGridChart(chartData, graphQuery)
+    const [reshapedData, rowKeys, rowLabels] = prepareDataForGridChart(chartData, graphQuery, intl.locale)
     return [reshapedData, rowKeys, rowLabels]
-  }, [data, query, name, testGroup])
+  }, [data, query, name, testGroup, intl])
 
   const headerOptions = { probe_cc: false, subtitle: false }
 
@@ -64,10 +64,10 @@ const Chart = React.memo(function Chart({testName, testGroup = null, title, quer
         <Box><Heading h={3} mt={40} mb={20}>{title}</Heading></Box>
         <Box>
           {(!chartData && !error) ? (
-            <div> Loading ...</div>
+            <div>{intl.formatMessage({id: 'General.Loading'})}</div>
           ) : (
             chartData === null || chartData.length === 0 ? (
-              <Heading h={5}>No Data</Heading>
+              <Heading h={5}>{intl.formatMessage({id: 'General.NoData'})}</Heading>
             ) : (
               <GridChart
                 data={chartData}
@@ -82,7 +82,7 @@ const Chart = React.memo(function Chart({testName, testGroup = null, title, quer
         {error &&
           <DetailsBox collapsed={false} content={<>
             <details>
-              <summary><span>Error: {error.message}</span></summary>
+              <summary><span>{intl.formatMessage({id: 'General.Error'})}: {error.message}</span></summary>
               <Box as='pre'>
                 {JSON.stringify(error, null, 2)}
               </Box>
@@ -95,4 +95,4 @@ const Chart = React.memo(function Chart({testName, testGroup = null, title, quer
   )
 })
 
-export default Chart
+export default ChartNetwork
